requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
beautifulsoup4==4.12.2
bs4==0.0.1
Pillow==10.1.0
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.16.0
tweepy==4.14.0